requirements.txt
fastapi
PyNaCl
requests